feat: Add mention handler module
The code changes include adding a new module called `mention_handler` in the `ui_tools` module. This module is responsible for handling mentions in the UI.

still a ton to do with this... but done for the night.
also store users are no longer Option<Vec<User>>, its just Vec<User>