added data table to mastertech for stock
also made sure that searching for stock works with lower case and you do
NOT have to type the brackets in for the item codes to search properly.
you can now download mastertech from the website without logging in.
added SEB search tool to mastertech (not fully done yet)
sorting by Serial number works better, if the text is 'Not Attached'
then ordering becomes std::cmp::Ordering::Less to 'not be included' in
the results of the sort, meaning it will stay at the bottom of the list
regardless.
fixed logging in mastertech to not include WGPU verbose garbage