a ton of bullshit.... - task audit table now has working notes upon opening the order, better view / layout, etc. - notes will auto add to thte database now as well even though there are no tasks associated - you can create a task from a service order from a button in the service order side panel - worked a lot with helper traits and fixing up the logic to allow all of the above
TODO: I need to give the task_id a value regardless of whether a task has been created or not, and make sure that ALL OF THE POSSIBLE LOCATIONS FOR CREATING A TASK USE THE SAME, CONSISTENT ID, because currently there is no way for me to find a task note without the task_id being present, if i do not have a cust thread or cust message id to search with. this is so i can prevent calling to prestashop every single time i want to pull notes from an order.
OR i can add service_number to task_note table as well..